% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/savis.R
\name{adaDimPlot}
\alias{adaDimPlot}
\title{adaDimPlot}
\usage{
adaDimPlot(umap_embedding, label, pt.size = 0, show.legend = TRUE)
}
\arguments{
\item{expr_matrix}{character}
}
\value{
nothing useful
}
\description{
Adaptively Plot the UMAP Embedding
}
\details{
Some details
}
\examples{
a<-1



}
