% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunSummary.R
\name{adaDimPlot2}
\alias{adaDimPlot2}
\title{adaDimPlot2}
\usage{
adaDimPlot2(umap_embedding, label, pt.size = 0)
}
\arguments{
\item{expr_matrix}{character}
}
\value{
nothing useful
}
\description{
Adaptively Plot the UMAP Embedding
}
\details{
Some details
}
\examples{
a<-1



}
