% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunSummary.R
\name{AdaptiveCombine}
\alias{AdaptiveCombine}
\title{AdaptiveCombine}
\usage{
AdaptiveCombine(
  expr_matrix,
  combined_embedding,
  cluster_label,
  cluster_label_i,
  npcs = 20,
  nfeatures = 2000,
  scale_factor_separation = 3,
  process_min_size = 0
)
}
\arguments{
\item{expr_matrix}{character}
}
\value{
nothing useful
}
\description{
Adaptively Combine PC and subPCs
}
\details{
Some details
}
\examples{
a<-1



}
