% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunSummary.R
\name{adjustUMAP_via_umap}
\alias{adjustUMAP_via_umap}
\title{ScaleFactor}
\usage{
adjustUMAP_via_umap(
  pca_embedding,
  umap_embedding,
  cluster_label,
  global_umap_embedding = NULL,
  distance_metric = "euclidean",
  scale_factor = 0.9,
  rotate = TRUE,
  density_adjust = TRUE,
  seed.use = 42,
  min_size = 100,
  maxit_push = NULL
)
}
\description{
Combined PC embedding with scale factor for subPC
}
\examples{
a<-1


}
