% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/savis.R
\name{ARIEvaluate}
\alias{ARIEvaluate}
\title{ARIEvaluate}
\usage{
ARIEvaluate(
  test_data,
  label,
  method = "louvain",
  resolution = 0.5,
  fixk = NULL
)
}
\arguments{
\item{expr_matrix}{character}
}
\value{
nothing useful
}
\description{
Adaptively Plot the UMAP Embedding
}
\details{
Some details
}
\examples{
a<-1





}
