% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunSummary.R
\name{CombinePC}
\alias{CombinePC}
\title{CombinePC}
\usage{
CombinePC(PC_embedding, cluster_label, sub_PC_list, scale_factor = NULL)
}
\arguments{
\item{expr_matrix}{character}
}
\value{
nothing useful
}
\description{
Combine PC and subPCs
}
\details{
Some details
}
\examples{
a<-1


}
