% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunSummary.R
\name{DimPlot1}
\alias{DimPlot1}
\title{DimPlot1}
\usage{
DimPlot1(
  embedding,
  group.by,
  axis.name = "SAVIS",
  pt.size = 0,
  show.legend = TRUE,
  seed.use = 42,
  text = TRUE,
  color.mode = 1
)
}
\arguments{
\item{expr_matrix}{character}
}
\value{
nothing useful
}
\description{
Adaptively Plot the UMAP Embedding
}
\details{
Some details
}
\examples{
a<-1



}
