% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunSummary.R
\name{DimPlot2}
\alias{DimPlot2}
\title{DimPlot2}
\usage{
DimPlot2(
  embedding,
  group.by,
  axis.name = "SAVIS",
  pt.size = 0,
  show.legend = TRUE,
  scale_color = c("grey", "blue")
)
}
\arguments{
\item{expr_matrix}{character}
}
\value{
nothing useful
}
\description{
Adaptively Plot the UMAP Embedding
}
\details{
Some details
}
\examples{
a<-1



}
