% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunSummary.R
\name{DoCluster}
\alias{DoCluster}
\title{DoCluster}
\usage{
DoCluster(
  pc_embedding,
  method = "louvain",
  resolution = 0.5,
  subcluster = FALSE,
  fixk = NULL,
  seed.use = 42,
  verbose = FALSE,
  verbose_more = FALSE
)
}
\arguments{
\item{expr_matrix}{character}
}
\value{
nothing useful
}
\description{
Do cluster for PC embedding
}
\details{
Some details
}
\examples{
a<-1


}
