% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunSummary.R
\name{FormAdaptiveCombineList}
\alias{FormAdaptiveCombineList}
\title{FormAdaptiveCombineList}
\usage{
FormAdaptiveCombineList(
  expr_matrix,
  expr_matrix_pca,
  max_stratification,
  stratification_count,
  scale_factor_separation,
  resolution,
  cluster_method,
  npcs,
  nfeatures,
  process_min_size,
  hvg_method = NULL,
  differentail_gene_cutoff = 20,
  do_cluster = TRUE,
  cluster_label = NULL,
  check_differential = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{expr_matrix}{character}
}
\value{
nothing useful
}
\description{
FormAdaptiveCombineList
}
\details{
Some details
}
\examples{
a<-1




}
