% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunSummary.R
\name{get_umap_embedding_adjust_umap}
\alias{get_umap_embedding_adjust_umap}
\title{ScaleFactor}
\usage{
get_umap_embedding_adjust_umap(
  pca_embedding,
  pca_center,
  pca_anchor_index,
  pca_dist,
  umap_embedding,
  N_label_,
  cluster_,
  label_index_,
  main_index = NULL,
  pca_dist_main = NULL,
  distance_metric = "euclidean",
  scale_factor = 0.9,
  rotate = TRUE,
  seed.use = 42
)
}
\description{
Combined PC embedding with scale factor for subPC
}
\examples{
a<-1


}
