% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunSummary.R
\name{RunAdaUMAP}
\alias{RunAdaUMAP}
\title{RunAdaUMAP}
\usage{
RunAdaUMAP(
  X,
  metric = "euclidean",
  metric_count = 1,
  py_envir = globalenv(),
  n.neighbors = 30L,
  n.components = 2L,
  n.epochs = NULL,
  learning.rate = 1,
  min.dist = 0.3,
  spread = 1,
  set.op.mix.ratio = 1,
  local.connectivity = 1L,
  repulsion.strength = 1,
  negative.sample.rate = 5,
  a = 1.8956,
  b = 0.8006,
  uwot.sgd = FALSE,
  seed.use = 42L,
  metric.kwds = NULL,
  angular.rp.forest = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{X}{sds}

\item{metric}{= 'euclidean',}

\item{metric_count}{= 1,}

\item{py_envir}{= globalenv(),}

\item{n.neighbors}{= 30L,}

\item{n.components}{= 2L,}

\item{n.epochs}{= NULL,}

\item{learning.rate}{= 1.0,}

\item{min.dist}{= 0.3,}

\item{spread}{= 1.0,}

\item{set.op.mix.ratio}{= 1.0,}

\item{local.connectivity}{= 1L,}

\item{repulsion.strength}{= 1,}

\item{negative.sample.rate}{= 5,}

\item{a}{= 1.8956,}

\item{b}{= 0.8006,}

\item{uwot.sgd}{= FALSE,}

\item{seed.use}{= 42L,}

\item{metric.kwds}{= NULL,}

\item{angular.rp.forest}{= FALSE,}

\item{verbose}{= FALSE}
}
\value{
nothing useful
}
\description{
Use Adaptive Distance Metric to Run UMAP
}
\details{
This function argument to the function
}
\examples{
a<-1

}
