% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunSummary.R
\name{RunPreSAVIS}
\alias{RunPreSAVIS}
\title{RunPreSAVIS}
\usage{
RunPreSAVIS(
  object,
  hvg_method = NULL,
  distance_metric = "euclidean",
  cluster_method = "louvain",
  resolution = 0.1,
  resolution_sub = 0,
  adaptive = TRUE,
  max_stratification = 3,
  scale_factor_separation = 3,
  process_min_size = NULL,
  process_min_count = NULL,
  check_differential = FALSE,
  verbose = TRUE,
  show_cluster = FALSE,
  return_cluster = FALSE,
  verbose_more = FALSE
)
}
\arguments{
\item{object}{sdsd}

\item{distance_metric}{The default is "euclidean".}

\item{cluster_method}{The default is "louvain". User can choose from c("louvain","spectral"). But "louvain" performs much better.}

\item{resolution}{The resolution for The default is 0.5.}

\item{resolution_sub}{The default is 0.}

\item{max_stratification}{The default is 3.}

\item{scale_factor_separation}{The default is 3.}

\item{process_min_size}{The default is NULL.}

\item{process_min_count}{The default is NULL.}

\item{check_differential}{The default is FALSE.}

\item{verbose}{The default is TRUE.}

\item{show_cluster}{The default is FALSE.}

\item{return_cluster}{The default is FALSE.}

\item{verbose_more}{The default is FALSE.}

\item{assay_for_var_features}{sds}
}
\value{
nothing useful
}
\description{
savis: single-cell RNAseq adaptive visualiztaion
}
\details{
This function argument to the function
}
\examples{
a<-1

}
