% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunSummary.R
\name{RunSAVIS}
\alias{RunSAVIS}
\title{RunSAVIS}
\usage{
RunSAVIS(
  object,
  adjust_UMAP = TRUE,
  adjust_method = "all",
  adjust_rotate = TRUE,
  shrink_distance = TRUE,
  density_adjust = TRUE,
  adjust_scale_factor = 0.9,
  verbose = TRUE,
  seed.use = 42L
)
}
\arguments{
\item{object}{sds}

\item{adjust_UMAP}{= TRUE,}

\item{adjust_method}{= "all",}

\item{adjust_rotate}{= TRUE,}

\item{shrink_distance}{= TRUE,}

\item{density_adjust}{= TRUE,}

\item{verbose}{= TRUE,}

\item{seed.use}{= 42L}
}
\value{
nothing useful
}
\description{
RunSAVIS
}
\details{
This function argument to the function
}
\examples{
a<-1

}
