% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunSummary.R
\name{savis}
\alias{savis}
\title{savis}
\usage{
savis(
  expr_matrix,
  is_count_matrix = TRUE,
  assay_for_var_features = "rawcount",
  npcs = 20,
  nfeatures = 2000,
  distance_metric = "euclidean",
  cluster_method = "louvain",
  resolution = 0.1,
  resolution_sub = 0,
  adaptive = TRUE,
  max_stratification = 3,
  scale_factor_separation = 3,
  process_min_size = NULL,
  process_min_count = NULL,
  run_adaUMAP = TRUE,
  adjust_UMAP = TRUE,
  adjust_method = "all",
  adjust_rotate = TRUE,
  shrink_distance = TRUE,
  density_adjust = TRUE,
  density_adjust_via_global_umap = FALSE,
  global_umap_embedding = NULL,
  check_differential = FALSE,
  verbose = TRUE,
  show_cluster = FALSE,
  return_cluster = FALSE,
  return_combined_PC = FALSE,
  verbose_more = FALSE,
  compressed_storage = TRUE,
  seed.use = 42L
)
}
\arguments{
\item{expr_matrix}{The expression matrix: gene(feature) as row; cell(sample) as column.}

\item{is_count_matrix}{Whether expr_matrix is count matrix or normalized version. If the expression matrix is count matrix, normalization will be performed. Default is TRUE.}

\item{assay_for_var_features}{The assay we use to select highly variable genes. Select from c("rawcount","normalizedcount"). Default is "rawcount".}

\item{npcs}{The number of principle components will be computed as the initialization input of nonlinear low dimensional embeddings. Default is 20.}

\item{nfeatures}{The number of highly variable genes will be selected. Default is 2000.}

\item{distance_metric}{The default is "euclidean". Recommend to use "euclidean" because we need to distinguish between global distance and local distance.}

\item{cluster_method}{The default is "louvain". User can choose from c("louvain","spectral"). But "louvain" performs much better.}

\item{resolution}{The resolution for the louvain clustering. The resolution ranges from 0 to 1. The lower resolution means conservative clustering(smaller number of clusters), while the higher resolution means aggressive clustering. The default is 0.5.}

\item{resolution_sub}{The resolution for the louvain clustering within subclusters, which means after the first step clustering and separation, we perform further clustering. The default is 0 because we are mainly interested in one step separation.}

\item{adaptive}{Whether we will run adaptive visualization. If adaptive is FALSE, we are just doing UMAP. The default is TRUE.}

\item{max_stratification}{The maximum level of stratification for subclustering. Set the maximum level to restrict too detailed exploration. The default is 3.}

\item{scale_factor_separation}{Scale factor used to distinguish global distance and local distance. The default is 3.}

\item{process_min_size}{The smallest size of cluster which we use to stop the process of subcluster evaluation. The clusters whose sizes are less than the cutoff will be not be further explored. The default is NULL.}

\item{process_min_count}{The processed clusters are determined by the ranking of the size of clusters. Only further explore the clusters whose rankings are less than the process_min_count. The default is NULL.}

\item{run_adaUMAP}{Whether we run the adaptive visualization. If the criterion is The default is TRUE.}

\item{adjust_UMAP}{The default is TRUE.}

\item{adjust_method}{The default is "all". Select from c("umap","mds").}

\item{adjust_rotate}{The default is TRUE.}

\item{shrink_distance}{The default is TRUE.}

\item{check_differential}{The default is FALSE.}

\item{verbose}{The default is TRUE.}

\item{show_cluster}{The default is FALSE.}

\item{return_cluster}{The default is FALSE.}

\item{verbose_more}{The default is FALSE.}

\item{seed.use}{The default is 42L}
}
\value{
nothing useful
}
\description{
savis: single-cell RNAseq adaptive visualiztaion
}
\details{
This function argument to the function
}
\examples{
a<-1

}
