% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunSummary.R
\name{ScaleFactor}
\alias{ScaleFactor}
\title{ScaleFactor}
\usage{
ScaleFactor(
  combined_embedding,
  cluster_label,
  npcs = NULL,
  center_method = "mean",
  scale_factor_separation = 3,
  return_factor = TRUE
)
}
\arguments{
\item{expr_matrix}{character}
}
\value{
nothing useful
}
\description{
Combined PC embedding with scale factor for subPC
}
\details{
Adapts source code from: https://github.com/rstudio/reticulate/blob/master/R/source.R

Now source_python_ supports loading from http
}
\examples{
a<-1


}
