% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunSummary.R
\name{SeuratLPCA}
\alias{SeuratLPCA}
\title{SeuratLPCA}
\usage{
SeuratLPCA(expr_matrix, method.names = NULL, npcs = 20, nfeatures = 2000)
}
\arguments{
\item{expr_matrix}{character}
}
\value{
nothing useful
}
\description{
Get PC matrix of single cell RNAseq matrix using limited memory with Seurat pipline
}
\details{
Some details
}
\examples{
a<-1



}
