% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunSummary.R
\name{adjustUMAP}
\alias{adjustUMAP}
\title{adjustUMAP}
\usage{
adjustUMAP(
  pca_embedding,
  umap_embedding,
  cluster_label = NULL,
  global_umap_embedding = NULL,
  adjust_method = "umap",
  distance_metric = "euclidean",
  scale_factor = 0.9,
  rotate = TRUE,
  density_adjust = TRUE,
  shrink_distance = TRUE,
  shrink_factor = 0.2,
  seed.use = 42,
  min_size = 100,
  maxit_push = NULL
)
}
\arguments{
\item{expr_matrix}{character}
}
\value{
nothing useful
}
\description{
Adjust UMAP to deal with distortion
}
\details{
amazing umap
}
\examples{
a<-1


}
