% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunSummary.R
\name{FormCombinedEmbedding}
\alias{FormCombinedEmbedding}
\title{FormCombinedEmbedding}
\usage{
FormCombinedEmbedding(
  expr_matrix,
  expr_matrix_pca,
  cluster_label,
  npcs,
  nfeatures,
  hvg_method = NULL,
  center_method = "mean",
  scale_factor_separation = 3
)
}
\arguments{
\item{expr_matrix}{character}
}
\value{
nothing useful
}
\description{
FormCombinedEmbedding
}
\details{
Some details
}
\examples{
a<-1



}
