% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/savis.R
\name{RunAdaUMAP}
\alias{RunAdaUMAP}
\title{RunAdaUMAP}
\usage{
RunAdaUMAP(
  X,
  metric = "euclidean",
  metric_count = 1,
  py_envir = parent.frame(),
  n.neighbors = 30L,
  n.components = 2L,
  n.epochs = NULL,
  learning.rate = 1,
  min.dist = 0.3,
  spread = 1,
  set.op.mix.ratio = 1,
  local.connectivity = 1L,
  repulsion.strength = 1,
  negative.sample.rate = 5,
  a = 1.8956,
  b = 0.8006,
  uwot.sgd = FALSE,
  seed.use = 42L,
  metric.kwds = NULL,
  angular.rp.forest = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{expr_matrix}{character}
}
\value{
nothing useful
}
\description{
Use Adaptive Distance Metric to Run UMAP
}
\details{
This function argument to the function
}
\examples{
a<-1

}
