% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunSummary.R
\name{savis}
\alias{savis}
\title{savis}
\usage{
savis(
  expr_matrix,
  is_count_matrix = TRUE,
  assay_for_var_features = "rawcount",
  npcs = 20,
  nfeatures = 2000,
  distance_metric = "euclidean",
  cluster_method = "louvain",
  resolution = 0.5,
  resolution_sub = 0,
  memory_save = FALSE,
  adaptive = TRUE,
  max_stratification = 3,
  scale_factor_separation = 3,
  process_min_size = NULL,
  process_min_count = NULL,
  run_adaUMAP = TRUE,
  adjust_UMAP = TRUE,
  adjust_method = "all",
  adjust_rotate = TRUE,
  shrink_distance = TRUE,
  density_adjust = TRUE,
  density_adjust_via_global_umap = FALSE,
  global_umap_embedding = NULL,
  check_differential = FALSE,
  verbose = TRUE,
  show_cluster = FALSE,
  return_cluster = FALSE,
  return_combined_PC = FALSE,
  verbose_more = FALSE,
  compressed_storage = TRUE,
  seed.use = 42L
)
}
\arguments{
\item{expr_matrix}{The expression matrix: gene(feature) as row; cell(sample) as column.}

\item{is_count_matrix}{Whether expr_matrix is count matrix or normalized version. If the expression matrix is count matrix, normalization will be performed. Default is TRUE.}

\item{assay_for_var_features}{}

\item{npcs}{The number of principle components will be computed. Default is 20.}

\item{nfeatures}{The number of highly variable genes will be selected. Default is 2000.}

\item{distance_metric}{The default is "euclidean".}

\item{cluster_method}{The default is "louvain". User can choose from c("louvain","spectral"). But "louvain" performs much better.}

\item{resolution}{The resolution for The default is 0.5.}

\item{resolution_sub}{The default is 0.}

\item{memory_save}{The default is FALSE. This function will take some storage to temporarily save the data from memory. Don't worry. SAVIS will soon delete it!!! Also, SAVIS uses unique name for the temporary data to keep everything safe!!!}

\item{adaptive}{The default is FALSE.}

\item{max_stratification}{The default is 3.}

\item{scale_factor_separation}{The default is 3.}

\item{process_min_size}{The default is NULL.}

\item{process_min_count}{The default is NULL.}

\item{run_adaUMAP}{The default is TRUE.}

\item{adjust_UMAP}{The default is TRUE.}

\item{adjust_method}{The default is "all". Select from c("umap","mds").}

\item{adjust_rotate}{The default is TRUE.}

\item{shrink_distance}{The default is TRUE.}

\item{check_differential}{The default is FALSE.}

\item{verbose}{The default is TRUE.}

\item{show_cluster}{The default is FALSE.}

\item{return_cluster}{The default is FALSE.}

\item{verbose_more}{The default is FALSE.}

\item{seed.use}{The default is 42L}
}
\value{
nothing useful
}
\description{
savis: single-cell RNAseq adaptive visualiztaion
}
\details{
This function argument to the function
}
\examples{
a<-1

}
