% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/savis.R
\name{savis}
\alias{savis}
\title{savis}
\usage{
savis(
  expr_matrix,
  is_count_matrix = TRUE,
  npcs = 10,
  nfeatures = 2000,
  distance_metric = "euclidean",
  cluster_method = "louvain",
  resolution = 0.5,
  resolution_sub = 0,
  adaptive = FALSE,
  max_stratification = 3,
  scale_factor_separation = 3,
  process_min_size = NULL,
  process_min_count = NULL,
  run_adaUMAP = TRUE,
  adjust_UMAP = TRUE,
  adjust_method = "all",
  adjust_rotate = TRUE,
  shrink_distance = TRUE,
  check_differential = FALSE,
  verbose = TRUE,
  show_cluster = FALSE,
  return_cluster = FALSE,
  return_global_umap = FALSE,
  verbose_more = FALSE,
  seed.use = 42L
)
}
\arguments{
\item{expr_matrix}{The expression matrix: gene(feature) as row; cell(sample) as column.}

\item{is_count_matrix}{Whether expr_matrix is count matrix or normalized version. If the expression matrix is count matrix, normalization will be performed. Default is TRUE.}

\item{npcs}{The number of principle components will be computed. Default is 10.}

\item{nfeatures}{The number of highly variable genes will be selected. Default is 2000.}

\item{distance_metric}{The default is "euclidean".}

\item{cluster_method}{The default is "louvain".}

\item{resolution}{The default is 0.5.}

\item{resolution_sub}{The default is 0.}

\item{adaptive}{The default is FALSE.}

\item{max_stratification}{The default is 3.}

\item{scale_factor_separation}{The default is3.}

\item{process_min_size}{The default is NULL.}

\item{process_min_count}{The default is NULL.}

\item{run_adaUMAP}{The default is TRUE.}

\item{adjust_UMAP}{The default is TRUE.}

\item{adjust_method}{The default is "all".}

\item{adjust_rotate}{The default is TRUE.}

\item{shrink_distance}{The default is TRUE.}

\item{check_differential}{The default is FALSE.}

\item{verbose}{The default is TRUE.}

\item{show_cluster}{The default is FALSE.}

\item{return_cluster}{The default is FALSE.}

\item{return_global_umap}{The default is FALSE.}

\item{verbose_more}{The default is FALSE.}

\item{seed.use}{The default is 42L}
}
\value{
nothing useful
}
\description{
savis: single-cell RNAseq adaptive visualiztaion
}
\details{
This function argument to the function
}
\examples{
a<-1

}
