% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunSummary.R
\name{SubPCEmbedding}
\alias{SubPCEmbedding}
\title{SubPCEmbedding}
\usage{
SubPCEmbedding(
  expr_matrix,
  cluster_label,
  npcs,
  nfeatures,
  hvg_method = NULL,
  return_hvg = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{expr_matrix}{character}
}
\value{
nothing useful
}
\description{
Based on clustering, calculate the subPC embedding for each cluster
}
\details{
Some details
}
\examples{
a<-1


}
